backend/python-service/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
pytube==15.0.0
openai==1.3.0
whisper==1.1.10
pydantic==2.4.0
python-dotenv==1.0.0 